pyproject.toml
[tool.poetry]
name = "wahti"
version = "0.1.0"
description = ""
authors = ["paapoh"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.10.0"
requests = "^2.27.1"
python-telegram-bot = "^13.11"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
